refactor names to be more accurate
