+++ web-ui/src/components/countryInfo/index.jsx
@@ -1,0 +1,3 @@
+import CountryInfo from "./countryInfo";
+
+export default CountryInfo;
