+++ web-ui/src/components/gradientInfo/gradientInfo.module.css
