+++ web-ui/src/components/gradientInfo/index.jsx
@@ -1,0 +1,3 @@
+import GradientInfo from "./gradientInfo";
+
+export default GradientInfo;
