--- web-ui/src/components/info/index.jsx
@@ -1,3 +1,0 @@
-import Info from "./info";
-
-export default Info;
