--- web-ui/src/components/legend/index.jsx
@@ -1,3 +1,0 @@
-import Legend from "./legend";
-
-export default Legend;
